miniblank was incorrectly uint8 samples while DMA was expecting half words

+++ sources/Adapters/adv/audio/advAudioDriver.h
@@ -11,7 +11,7 @@
 #include "Foundation/T_Singleton.h"
 #include "Services/Audio/AudioDriver.h"
 
-#define MINI_BLANK_SIZE 256 // Samples
+#define MINI_BLANK_SIZE 256 // 16-bit samples
 
 class advAudioDriver : public AudioDriver {
 public:
@@ -39,7 +39,7 @@ private:
   static advAudioDriver *instance_;
 
   AudioSettings settings_;
-  static const uint8_t miniBlank_[MINI_BLANK_SIZE];
+  static const int16_t miniBlank_[MINI_BLANK_SIZE];
 
   uint32_t startTime_;
 };
